Use conditions on auto-configuration
This commit improves our auto-configuration to be smarter about the
environment.

First, `HelloService` might not be available so we shouldn't be trying
to configure it if that's the case.

Then, if the user has configured their own `HelloService` we shouldn't
override that. This showcases that user configuration is processed first
to get a chance for that condition to apply properly.

Now that our auto-configuration is fixed, we can remove our own
override and let the default apply. Unfortunately, doing so has changed
the message, from `Howdy World!` to `Hello World!`. Let's fix that in
the next commit.